requirements.txt
Django==1.11
django-location-field==2.0.3
olefile==0.44
Pillow==2.5.0
pytz==2017.2
six==1.11.0
